moved itest to its own package
remove ^M
