+++ src/test/java/org/springframework/data/crate/integration/ValidatingCrateEventListenerIntegrationTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.data.crate.core.mapping.event;
+package org.springframework.data.crate.integration;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.junit.Assert.assertThat;
@@ -27,7 +27,8 @@ import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.data.crate.CrateIntegrationTest;
+import org.springframework.data.crate.core.mapping.event.LifecycleEventConfigurationBase;
+import org.springframework.data.crate.core.mapping.event.User;
 import org.springframework.data.crate.core.CrateOperations;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
@@ -40,8 +41,8 @@ import java.util.Locale;
  * @since 1.0.0
  */
 @RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(classes={ValidatingCrateEventListenerTest.TestConfiguration.class})
-public class ValidatingCrateEventListenerTest extends CrateIntegrationTest {
+@ContextConfiguration(classes={ValidatingCrateEventListenerIntegrationTest.TestConfiguration.class})
+public class ValidatingCrateEventListenerIntegrationTest extends CrateBaseIntegrationTest {
 
 	@Autowired
 	CrateOperations crateOperations;
